Rename factory.py to factories.py
